Split Payloads into multiple files
