+++ common/src/commonMain/kotlin/com/studo/campusqr/common/payloads/LoginPayloads.kt
@@ -1,0 +1,10 @@
+package com.studo.campusqr.common.payloads
+
+import com.studo.campusqr.common.ClientPayload
+import kotlinx.serialization.Serializable
+
+@Serializable
+class MailLoginData(
+  val email: String,
+  val password: String
+) : ClientPayload
