+++ moderatorFrontend/src/main/kotlin/views/locations/locationsOverview/LocationTableRow.kt
@@ -3,6 +3,7 @@ package views.locations.locationsOverview
 import app.baseUrl
 import app.routeContext
 import com.studo.campusqr.common.*
+import com.studo.campusqr.common.payloads.LocationVisitData
 import kotlinx.browser.window
 import react.*
 import util.*
