+++ server/src/main/kotlin/com/studo/campusqr/endpoints/ReportData.kt
@@ -1,9 +1,10 @@
 package com.studo.campusqr.endpoints
 
 import com.moshbit.katerbase.*
-import com.studo.campusqr.common.ActiveCheckIn
-import com.studo.campusqr.common.ReportData
 import com.studo.campusqr.common.emailSeparators
+import com.studo.campusqr.common.payloads.ActiveCheckIn
+import com.studo.campusqr.common.payloads.ReportData
+import com.studo.campusqr.common.payloads.TraceContactsReportData
 import com.studo.campusqr.database.BackendLocation
 import com.studo.campusqr.database.BackendSeatFilter
 import com.studo.campusqr.database.CheckIn
@@ -226,11 +227,10 @@ suspend fun AuthenticatedApplicationCall.returnReportData() {
     return
   }
 
-  val params = receiveJsonStringMap()
+  val params: TraceContactsReportData = receiveClientPayload()
 
-  val now = Date()
-  val emails = params.getValue("email").split(*emailSeparators).filter { it.isNotEmpty() }
-  val oldestDate = params["oldestDate"]?.toLong()?.let { Date(it) } ?: now.addDays(-14)
+  val emails = params.email.split(*emailSeparators).filter { it.isNotEmpty() }
+  val oldestDate = params.oldestDate.toLong().let { Date(it) }
 
   respondObject(generateContactTracingReport(emails, oldestDate))
 }
